refactor: gate malleable code behind optional SharpireMalleable library
Make malleable profile support opt-in at compile time so plain-http
Sharpire builds don't carry the JSON parser, transform engine, or
System.Web.Extensions (JavaScriptSerializer) dependency.

Base Sharpire (Sharpire/) becomes malleable-agnostic:
  - SessionInfo, Coms, EmpireStager → partial classes
  - Empire.Agent.MalleableProfile.cs and .MalleableTransform.cs moved
    out of the base library
  - Malleable hook points replaced by partial-void method stubs
    (TryMalleableGet, TryMalleableSend, TryPickMalleableStagerUri);
    when the extension isn't compiled in they compile to no-ops and
    the caller falls through to the legacy cookie/task-URI path

New SharpireMalleable/ sibling library holds everything malleable:
  - The moved MalleableProfile + MalleableTransform classes
  - Empire.Agent.Malleable.cs carrying three partial-class extensions
    that fill in the hooks and provide GetMalleableProfile /
    SetMalleableProfile on SessionInfo

When Empire's http_malleable listener generates a C# stager it lists
BOTH Sharpire and SharpireMalleable as ReferenceSourceLibraries so
EmpireCompiler globs both directories' .cs files into one assembly.
The plain http listener only lists Sharpire — the malleable types
simply don't exist in that binary.
