+++ Sharpire/Sharpire.csproj
@@ -61,8 +61,6 @@
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
-    <Reference Include="System.Web" />
-    <Reference Include="System.Web.Extensions" />
   </ItemGroup>
   <ItemGroup>
   <Reference Include="System.Management.Automation">
@@ -78,8 +76,6 @@
     <Compile Include="Empire.Agent.Jobs.cs" />
     <Compile Include="Empire.Agent.Stager.cs" />
     <Compile Include="Empire.Agent.AesGcm.cs" />
-    <Compile Include="Empire.Agent.MalleableProfile.cs" />
-    <Compile Include="Empire.Agent.MalleableTransform.cs" />
     <Compile Include="Misc.cs" />
     <Compile Include="Program.cs" />
   </ItemGroup>
