+++ SharpireMalleable/SharpireMalleable.csproj
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  SharpireMalleable is an optional source-library extension to Sharpire.
+  When Empire's malleable HTTP listener generates a C# launcher it lists
+  BOTH Sharpire and SharpireMalleable as ReferenceSourceLibraries so
+  EmpireCompiler globs both directories' .cs files into one compilation
+  unit. The partial classes in Empire.Agent.Malleable.cs then fill in the
+  partial-method hooks declared in base Sharpire.
+
+  The plain http listener only lists Sharpire, so base binaries ship
+  without the malleable types or the JavaScriptSerializer dependency on
+  System.Web.Extensions.
+
+  This csproj is provided for IDE convenience only — EmpireCompiler does
+  not consume .csproj <Compile> entries; it recursively globs *.cs from
+  the ReferenceSourceLibrary's Location directory.
+-->
+<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <ProjectGuid>{2A8F47C1-3B5E-4D2A-9F1C-BE9CF6A2D813}</ProjectGuid>
+    <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
+    <TargetFrameworkProfile />
+    <LangVersion>7</LangVersion>
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Web" />
+    <Reference Include="System.Web.Extensions" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="Empire.Agent.MalleableProfile.cs" />
+    <Compile Include="Empire.Agent.MalleableTransform.cs" />
+    <Compile Include="Empire.Agent.Malleable.cs" />
+    <!-- Base-library .cs files are listed via Link so the IDE sees them
+         alongside the extension. EmpireCompiler ignores this list. -->
+    <Compile Include="..\Sharpire\Empire.Agent.cs">
+      <Link>Base\Empire.Agent.cs</Link>
+    </Compile>
+    <Compile Include="..\Sharpire\Empire.Agent.Coms.cs">
+      <Link>Base\Empire.Agent.Coms.cs</Link>
+    </Compile>
+    <Compile Include="..\Sharpire\Empire.Agent.Stager.cs">
+      <Link>Base\Empire.Agent.Stager.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+</Project>
